requirements.txt
# Core Dependencies
Django==5.0.1
django-crispy-forms==2.1
django-environ==0.11.2
crispy-tailwind==0.5.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
python-docx==1.1.0

# ML and CUDA Dependencies
--find-links https://download.pytorch.org/whl/torch_stable.html
--extra-index-url https://pypi.org/simple
torch==2.1.2+cu118
torchvision==0.16.2+cu118
torchaudio==2.1.2+cu118
transformers==4.36.2
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scikit-learn>=1.0.0
scipy>=1.10.0

# Google AI Dependencies
google-generativeai
together

# Audio Processing
ffmpeg-python==0.2.0
soundfile==0.12.1
numba>=0.56.0
librosa

# OCR Dependencies
pytesseract==0.3.13
Pillow==10.4.0
pdf2image==1.17.0
markdown2
openpyxl>=3.1.2  # Required for pandas Excel support

# API and Web
fastapi==0.115.5
uvicorn==0.32.1
aiohttp==3.11.2
requests==2.32.3

# Utilities
python-dotenv==1.0.1
tqdm==4.66.4
beautifulsoup4==4.12.3
bs4==0.0.2
httpx==0.28.0
Jinja2==3.1.3
Markdown==3.7
PyYAML==6.0.2
spacy>=3.0.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl
